fix publish

+++ function.groovy
@@ -566,8 +566,14 @@ def set_npm_nexus_publish()
     npm config set email "${devops_email}"
     npm config set always-auth true
     npm config set _auth "${auth_nexus}"
-    mv ./package-common.json ./lib/package.json
-    mv ./.npmrc-common ./lib/.npmrc
+    if [ -e "package-common.json" ] ; then
+      mv ./package-common.json ./lib/package.json
+    fi
+    fi [ -e ".npmrc-common" ] ; then
+      mv ./.npmrc-common ./lib/.npmrc
+    fi
+    cp /home/jenkins/.npmrc .npmrc
+    cp /home/jenkins/.npmrc lib/
     cd lib/
     npm publish
   '''
